DataManager update
further memory optimisation
changed datamanager to a namespace
Added dataloader and datasaver
Improved card's scene
Cards can now change characters
Added a little enum so i change sprites
Little adjustments for card.cs

+++ Scripts/Editor/ComposerLib.cs
@@ -174,7 +174,7 @@ public partial class ComposerLib : Node
     void DeferredGotoScene(string scene, Dictionary<string,Variant> args = null)
     {
         bool _isAnimated = args.ContainsKey("IsAnimated") ? (bool)args["IsAnimated"] : false;
-        ANIMATIONS _animation = args.ContainsKey("Animation") ? (ANIMATIONS)((int)args["Animation"]) : ANIMATIONS.NONE;
+        ANIMATIONS _animation = args.ContainsKey("Animation") ? (ANIMATIONS)((byte)args["Animation"]) : ANIMATIONS.NONE;
 
         if (_isAnimated && _animation != ANIMATIONS.NONE)
         {
